gltfio: Add internal MorphHelper, enable up to 256 targets.
Previously, gltfio could not animate models with more than 4 morph
targets. Now we can handle up to 256 targets, as long as no more than
4 targets are ever used simultaneously. This allows us to handle
many more models, such as those exported by Modo, and the horse model
in #1852.

We do this by picking the highest 4 weights. The indices for these
weights are called the "primary indices", and we construct a new
VertexBuffer for each unique set of primary indices.

In future versions of Filament this could be made more efficient
by aggregating the buffers and moving a "base vertex" over the
course of the animation.

Fixes #1487.